Move positionwise_feed_forward layer and modify related code
