+++ attention/layers/transformer.py
@@ -1,5 +1,5 @@
 import torch.nn as nn
-from attention.sublayers.transformer import MultiHeadAttention, PositionwiseFeedForwrd
+from attention.sublayers.transformer import MultiHeadAttention, PositionwiseFeedForward
 
 
 class EncoderLayer(nn.Module):
@@ -19,7 +19,7 @@ class EncoderLayer(nn.Module):
             d_v=d_v,
             dropout=dropout
         )
-        self.pos_ffn = PositionwiseFeedForwrd(
+        self.pos_ffn = PositionwiseFeedForward(
             d_in=d_model,
             d_hid=d_inner,
             dropout=dropout
@@ -61,7 +61,7 @@ class DecoderLayer(nn.Module):
             d_v=d_v,
             dropout=dropout
         )
-        self.pos_ffn = PositionwiseFeedForwrd(
+        self.pos_ffn = PositionwiseFeedForward(
             d_in=d_model,
             d_hid=d_inner,
             dropout=dropout
